FEAT : 메세지 조회 기능 구현
미리 작성된 틀과 구조가 달라서, 틀은 그대로 남겨두고 새로 기능들을 구현했음

+++ v2/backend/greetingCard/src/main/java/com/cabi/greetingCard/dto/MessageResponsePaginationDto.java
@@ -9,5 +9,5 @@ import lombok.Getter;
 public class MessageResponsePaginationDto {
 
 	private List<MessageResponseDto> messages;
-	private Long totalLength;
+	private int totalLength;
 }
